+++ 3rdparty/libbacktrace/mmapio.c
@@ -67,7 +67,7 @@ backtrace_get_view (struct backtrace_state *state ATTRIBUTE_UNUSED,
     }
 
   pagesize = getpagesize ();
-  inpage = offset % pagesize;
+  inpage = (unsigned int)(offset % pagesize);
   pageoff = offset - inpage;
 
   size += inpage;
